Moved periodic block functions to qttools
- New `toeplitz submodule.
- `.shape[-1]` defined at the beginning
